FINAL PROJECT
Finally we're at the end of the course and to make sure that your
learned all the things needed to make new applications we decided to
gift your this sample but real world application to give a few ideas in
how a flask app looks like.

We hope you can enjoy it! :)

Last thing about it is that this project comes with a few errors to be
fixed, we know you can fix every single one of them because now you know
how to do it, right? Good luck!

@codenoschool